==> database_parsing/1_preprocess_foodb_data.ipynb <==
*MeNu GUIDE*

# Preprocessing FooDB Data

import pandas as pd
import re

## Load relevant data tables

foodb_folder = "/path/to/downloaded/FooDB/data/"
processed_data_folder = "/path/to/processed/data/folder/"

foodb_compounds = pd.read_csv(f"{foodb_folder}Compound.csv",
                              dtype={'moldb_iupac': 'string', 'state': 'string', 'annotation_quality': 'string', 'description': 'string', 'kingdom': 'string', 'superklass': 'string', 'klass': 'string', 'subklass': 'string'})

foodb_descriptors = pd.read_csv(f"{foodb_folder}CompoundExternalDescriptor.csv")

## Process compounds table

### Drop unneccessary columns

foodb_compounds = foodb_compounds.drop(columns=['moldb_iupac', 'state'])
foodb_compounds.head()

### Rename misnamed columns

# Seems like they mixed up the column names in the downloadable FooDB file, so this should sort it out and match the correct column name with the correct info
foodb_compounds = foodb_compounds.rename(columns={'annotation_quality': 'description', 'description': 'cas_number', 'cas_number': 'smiles', 'moldb_inchikey': 'inchi', 'moldb_inchi': 'mono_mass', 'moldb_smiles': 'inchikey', 'moldb_mono_mass': 'iupac', 'superklass': 'superclass', 'klass': 'class', 'subklass': 'subclass'})
foodb_compounds.head()

## Process external descriptor table

### Add database origin to external identifiers

In order to match the external descriptors to the original compounds data it will have to be transformed slightly, so that it is only one unique compound per row with its various identifiers (ChEBI, LipidMaps, etc.) as columns.

def match_descriptor_to_database(descriptor:str):
    if descriptor.startswith('CHEBI'):
        return 'chebi'
    elif descriptor.startswith(('LMPK', 'LMST', 'LMFA', 'LMGP', 'LMGL', 'LMPR', 'LMSP')):
        return 'lipid_maps'
    elif descriptor.startswith('CPD'):
        return 'meta_cyc'
    elif re.match(r'^C[0-9]{5}$', descriptor):
        return 'kegg'
    else:
        return 'synonym'

foodb_descriptors['database'] = foodb_descriptors.external_id.apply(match_descriptor_to_database)

# extract only relevant information as preparation to pivot table
foodb_descriptors = foodb_descriptors[['external_id', 'compound_id', 'database']]
foodb_descriptors.head()

### Deduplicate entries

# check for duplicates as pivot does not work otherwise
foodb_descriptors[foodb_descriptors.duplicated(subset=['compound_id', 'database'])]

# have a closer look at duplicates
foodb_descriptors[(foodb_descriptors.compound_id == 1161) | (foodb_descriptors.compound_id == 8291)]

# match them back to the original compounds table
foodb_compounds[(foodb_compounds.id == 1161) | (foodb_compounds.id == 8291)]

It seems like there is no match for compound_id 8291 surprisingly.

After having a closer look at the ChEBI website and the entries, we will deduplicate the table by keeping the first entry, as CHEBI:47962 seems to be the correct one.

foodb_descriptors_deduplicated = foodb_descriptors.drop_duplicates(subset=['compound_id', 'database'])

### Pivot table in preparation for merge

foodb_descriptors_deduplicated_pivot = foodb_descriptors_deduplicated.pivot(index='compound_id', columns='database', values='external_id')
foodb_descriptors_deduplicated_pivot

# reformat table so that the compound_id is a normal colum instead of an index for the merging process
foodb_descriptors_deduplicated_pivot = foodb_descriptors_deduplicated_pivot.reset_index()
foodb_descriptors_deduplicated_pivot = foodb_descriptors_deduplicated_pivot.rename(columns={'compound_id': 'id'})
foodb_descriptors_deduplicated_pivot

## Merging of dataframes

foodb_compounds_merged = foodb_compounds.merge(foodb_descriptors_deduplicated_pivot, on='id', how='left')
foodb_compounds_merged

foodb_compounds_merged.info()

foodb_compounds_merged.head()

foodb_compounds_merged.to_csv(f'{processed_data_folder}foodb_compounds_with_external_descriptors.csv', index=False)
