==> database_parsing/2_preprocess_hmdb_data.ipynb <==
*MeNu GUIDE*

# Preprocessing HMDB Data

import pandas as pd
import xml.etree.ElementTree as ET

hmdb_folder = "/path/to/downloaded/HMDB/data/"
processed_data_folder = "/path/to/processed/data/folder/"

## Load XML file
This can take a while, as the file is around 6GB

tree = ET.parse(f'{hmdb_folder}hmdb_metabolites.xml')
root = tree.getroot()

## Extract all metabolites and their properties

compounds = []

for child in root:
    child_dict = {}
    
    for elem in child:
        # Some texts are None, some just contain white space, both of which we want to sort out. 
        if elem.text:
            text = elem.text.strip()
            if text:
                tag = elem.tag.split('}')[1]
                child_dict[tag] = text
    
    compounds.append(child_dict)

compounds_df = pd.DataFrame(compounds)

compounds_df

compounds_df = compounds_df.drop(columns=['version', 'status', 'creation_date', 'update_date', 'average_molecular_weight', 'traditional_iupac', 'state', 'synthesis_reference'])

## Check for duplicates with the same ChEBI ID
HMDB has some incorrect, duplicate annotations for certain external identifiers, among them ChEBI. As we want to merge the compounds from HMDB with ChEBI, we need to take care of this issue. Unfortunately, this is something that needs to be done manually. It might also be, that there are even more falsely assigned external identifiers that do not appear multiple times, but there is no easy way to check for that, as we can't go manually through thousands of compounds, so that might be something we will have to check at a later timepoint.

duplicated_chebi_ids = list(compounds_df[(compounds_df.chebi_id.duplicated()) & (compounds_df.chebi_id.notna())].chebi_id)
len(duplicated_chebi_ids)

compounds_for_manual_inspection = compounds_df[compounds_df.chebi_id.isin(duplicated_chebi_ids)].sort_values(by='chebi_id')[['name', 'accession', 'chebi_id']]
compounds_for_manual_inspection.to_csv(f"{hmdb_folder}compounds_for_manual_chebi_id_inspection.csv")

$\rightarrow$ Manual inspection step inbetween

compounds_for_manual_inspection = pd.read_csv(f"{hmdb_folder}compounds_for_manual_chebi_id_inspection.csv", sep=';')
compounds_for_manual_inspection

### Integrate the manually corrected ChEBI IDs

compounds_for_manual_inspection.loc[:, 'website check'] = compounds_for_manual_inspection['website check'].apply(lambda x: x.split('CHEBI:')[1] if pd.notna(x) else float('NaN'))

compounds_for_manual_inspection.index = compounds_for_manual_inspection.hmdb_id
compounds_for_manual_inspection = compounds_for_manual_inspection['website check']
compounds_correct_chebi_dict = compounds_for_manual_inspection.to_dict()

columns_to_modify = ['chemspider_id', 'drugbank_id', 'foodb_id', 'pubchem_compound_id', 'knapsack_id', 'kegg_id', 'wikipedia_id', 'metlin_id', 'biocyc_id', 'bigg_id', 'vmh_id', 'phenol_explorer_compound_id', 'pdb_id']

adjusted_df = compounds_df.copy(deep=True)
adjusted_df = adjusted_df.sort_values(by='chebi_id')

for column in columns_to_modify:
    adjusted_df[column] = adjusted_df.apply(lambda row: row[column] if (row.accession not in compounds_correct_chebi_dict.keys()) else (row[column] if row.chebi_id == compounds_correct_chebi_dict[row.accession] else float('NaN')), axis=1)

adjusted_df['chebi_id'] = adjusted_df.apply(lambda row: row.chebi_id if (row.accession not in compounds_correct_chebi_dict.keys()) else compounds_correct_chebi_dict[row.accession], axis=1)

adjusted_df[(adjusted_df.chebi_id.duplicated()) & (adjusted_df.chebi_id.notna())][['accession', 'name', 'chebi_id']]

adjusted_df = adjusted_df[adjusted_df.accession != 'HMDB0059641']
adjusted_df = adjusted_df[adjusted_df.accession != 'HMDB0006769']
adjusted_df = adjusted_df[adjusted_df.accession != 'HMDB0006759']
adjusted_df = adjusted_df[adjusted_df.accession != 'HMDB0010375']

adjusted_df.to_csv(f'{processed_data_folder}hmdb_metabolites.csv', index=False)
