==> database_parsing/3_match_foodb_hmdb.ipynb <==
*MeNu GUIDE*

# Matching FooDB and HMDB Compounds

import pandas as pd

processed_data_folder = "/path/to/processed/data/folder/"

foodb_df = pd.read_csv(f'{processed_data_folder}foodb_compounds_with_external_descriptors.csv', dtype={'description': 'string', 'cas_number': 'string', 'kingdom': 'string', 'superclass': 'string', 'class': 'string', 'subclass': 'string', 'chebi': 'string', 'kegg': 'string','lipid_maps': 'string', 'meta_cyc': 'string', 'synonym': 'string'})

hmdb_df = pd.read_csv(f'{processed_data_folder}hmdb_metabolites.csv', dtype={'cas_registry_number': 'string', 'drugbank_id': 'string', 'knapsack_id': 'string', 'kegg_id': 'string', 'biocyc_id': 'string', 'vmh_id': 'string', 'pdb_id': 'string', 'chebi_id': 'string'})

## Merge FooDB and HMDB
### Rename identifier columns so they match.

hmdb_df.columns

foodb_df.columns

hmdb_df = hmdb_df.rename(columns={'accession': 'hmdb_id', 'monisotopic_molecular_weight': 'mono_mass', 'iupac_name': 'iupac', 'cas_registry_number': 'cas_number'}) 

foodb_df = foodb_df.rename(columns={'id': 'foodb_id_internal', 'public_id': 'foodb_id', 'chebi': 'chebi_id', 'kegg': 'kegg_id'})

## Add names in lowercase and check merging results for using only foodb_id, only the name or both

hmdb_df['name'] = hmdb_df['name'].str.lower()
foodb_df['name'] = foodb_df['name'].str.lower()

foodb_hmdb_merge = hmdb_df.merge(foodb_df, how='outer', on='foodb_id', suffixes=['_hmdb', '_foodb'])
foodb_hmdb_merge_with_name = hmdb_df.merge(foodb_df, how='outer', on=['foodb_id', 'name'], suffixes=['_hmdb', '_foodb'])
foodb_hmdb_merge_only_name = hmdb_df.merge(foodb_df, how='outer', on=['name'], suffixes=['_hmdb', '_foodb'])

### Merged Data Analysis
Have a closer look at final table to see how well the matching worked and if there are any duplicates.

print(f'Count of HMDB IDs: {foodb_hmdb_merge.hmdb_id.count()}')
print(f'Count of unique HMDB IDs: {foodb_hmdb_merge.hmdb_id.nunique()}\n')

print(f'Count of FooDB IDs: {foodb_hmdb_merge.foodb_id_internal.count()}')
print(f'Count of unique FooDB IDs: {foodb_hmdb_merge.foodb_id_internal.nunique()}\n')

print(f'Number of compounds that could be matched between FooDB and HMDB: {foodb_hmdb_merge[foodb_hmdb_merge.foodb_id_internal.notna()].hmdb_id.count()}')
print(f'Number of HMDB compounds that could not be matched to FooDB: {foodb_hmdb_merge[foodb_hmdb_merge.foodb_id_internal.isna()].hmdb_id.count()}')
print(f'Number of FooDB compounds that could not be matched to HMDB: {foodb_hmdb_merge[foodb_hmdb_merge.hmdb_id.isna()].foodb_id.count()}\n')

print(f'Count of duplicate matches: {foodb_hmdb_merge.foodb_id_internal.count()-foodb_hmdb_merge.foodb_id_internal.nunique()}')

print(f'Count of HMDB IDs: {foodb_hmdb_merge_with_name.hmdb_id.count()}')
print(f'Count of unique HMDB IDs: {foodb_hmdb_merge_with_name.hmdb_id.nunique()}\n')

print(f'Count of FooDB IDs: {foodb_hmdb_merge_with_name.foodb_id_internal.count()}')
print(f'Count of unique FooDB IDs: {foodb_hmdb_merge_with_name.foodb_id_internal.nunique()}\n')

print(f'Number of compounds that could be matched between FooDB and HMDB: {foodb_hmdb_merge_with_name[foodb_hmdb_merge_with_name.foodb_id_internal.notna()].hmdb_id.count()}')
print(f'Number of HMDB compounds that could not be matched to FooDB: {foodb_hmdb_merge_with_name[foodb_hmdb_merge_with_name.foodb_id_internal.isna()].hmdb_id.count()}')
print(f'Number of FooDB compounds that could not be matched to HMDB: {foodb_hmdb_merge_with_name[foodb_hmdb_merge_with_name.hmdb_id.isna()].foodb_id_internal.count()}\n')

print(f'Count of duplicate matches: {foodb_hmdb_merge_with_name.foodb_id_internal.count()-foodb_hmdb_merge_with_name.foodb_id_internal.nunique()}')

print(f'Count of HMDB IDs: {foodb_hmdb_merge_only_name.hmdb_id.count()}')
print(f'Count of unique HMDB IDs: {foodb_hmdb_merge_only_name.hmdb_id.nunique()}\n')

print(f'Count of FooDB IDs: {foodb_hmdb_merge_only_name.foodb_id_internal.count()}')
print(f'Count of unique FooDB IDs: {foodb_hmdb_merge_only_name.foodb_id_internal.nunique()}\n')

print(f'Number of compounds that could be matched between FooDB and HMDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.foodb_id_internal.notna()].hmdb_id.count()}')
print(f'Number of HMDB compounds that could not be matched to FooDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.foodb_id_internal.isna()].hmdb_id.count()}')
print(f'Number of FooDB compounds that could not be matched to HMDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.hmdb_id.isna()].foodb_id_internal.count()}\n')

print(f'Count of duplicate matches: {foodb_hmdb_merge_only_name.foodb_id_internal.count()-foodb_hmdb_merge_only_name.foodb_id_internal.nunique()}')

# Check out duplicate matches for merging only via foodb_id. Seems like maybe there sometimes is an incorrect foodb_id saved for some HMDB entries
duplicate_matches = list(foodb_hmdb_merge[(foodb_hmdb_merge.foodb_id_internal.notna()) & (foodb_hmdb_merge.duplicated(subset='foodb_id_internal'))]['foodb_id'].unique())

foodb_hmdb_merge[foodb_hmdb_merge.foodb_id.isin(duplicate_matches)][['name_hmdb', 'name_foodb', 'hmdb_id', 'foodb_id', 'chebi_id_hmdb', 'chebi_id_foodb']]

# Check if there are any inconsistencies in foodb_ids after matching only via name
foodb_hmdb_merge_only_name[(foodb_hmdb_merge_only_name.foodb_id_hmdb != foodb_hmdb_merge_only_name.foodb_id_foodb) & (foodb_hmdb_merge_only_name.foodb_id_hmdb.notna()) & (foodb_hmdb_merge_only_name.foodb_id_foodb.notna())][['name', 'foodb_id_foodb', 'foodb_id_hmdb']]

After comparing the entries on the respective websites, it seems it makes the most amount of sense to merge based on name. This yields the most hits, while also having no duplicates. Now we need to merge all the identifiers. It makes sense that the foodb_id should be preferred, but if there is none, we should tak the foodb_id provided by HMDB.

foodb_hmdb_merge_only_name.loc[:, 'foodb_id'] = foodb_hmdb_merge_only_name.apply(lambda row: row.foodb_id_foodb if pd.notna(row.foodb_id_foodb) else None, axis=1)
foodb_hmdb_merge_only_name = foodb_hmdb_merge_only_name.drop(columns=['foodb_id_hmdb', 'foodb_id_foodb'])

print(f'Count of HMDB IDs: {foodb_hmdb_merge_only_name.hmdb_id.count()}')
print(f'Count of unique HMDB IDs: {foodb_hmdb_merge_only_name.hmdb_id.nunique()}\n')

print(f'Count of FooDB IDs: {foodb_hmdb_merge_only_name.foodb_id.count()}')
print(f'Count of unique FooDB IDs: {foodb_hmdb_merge_only_name.foodb_id.nunique()}\n')

print(f'Number of compounds that could be matched between FooDB and HMDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.foodb_id_internal.notna()].hmdb_id.count()}')
print(f'Number of HMDB compounds that could not be matched to FooDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.foodb_id_internal.isna()].hmdb_id.count()}')
print(f'Number of FooDB compounds that could not be matched to HMDB: {foodb_hmdb_merge_only_name[foodb_hmdb_merge_only_name.hmdb_id.isna()].foodb_id_internal.count()}\n')

print(f'Count of duplicate matches: {foodb_hmdb_merge_only_name.foodb_id_internal.count()-foodb_hmdb_merge_only_name.foodb_id_internal.nunique()}')

## Clean up columns of merged data

foodb_hmdb_merge_only_name.columns

# Merge descriptions. Strategy: HMDB seems to be correct more often than FooDB, so go with HMDB info, unless there is only FooDB info available. 
columns_to_clean = ["description", "mono_mass", "iupac", "inchi", "inchikey", "cas_number", "smiles", "chebi_id", "kegg_id"]

for column in columns_to_clean:
    column_foodb = column + '_foodb'
    column_hmdb = column + '_hmdb'
    
    foodb_hmdb_merge_only_name[column] = foodb_hmdb_merge_only_name.apply(lambda row: row[column_hmdb] if pd.notna(row[column_hmdb]) else row[column_foodb], axis=1)
    foodb_hmdb_merge_only_name = foodb_hmdb_merge_only_name.drop(columns=[column_foodb, column_hmdb])
    
    if column == 'chebi_id':
        foodb_hmdb_merge_only_name[column] = foodb_hmdb_merge_only_name[column].apply(lambda x: x.strip('CHEBI:') if pd.notna(x) else x)

foodb_hmdb_merge_only_name.columns

foodb_hmdb_merge_only_name.to_csv(f'{processed_data_folder}/foodb_hmdb.csv', index=False)

## Conclusion:
* Match was most successful if using only the name cast to lower case
* Various problems are likely due to incorrect entries in HMDB and the fact that the FooDB version that can be downloaded from the website is not up to date, which is probably why HMDB has a higher count of unique FooDB IDs than FooDB
* The process has been adjusted so that there are no duplicate matches
