==> database_parsing/8_match_exposome_explorer.ipynb <==
*MeNu GUIDE*

# Match Exposome Explorer

import os.path
import pandas as pd

exposome_explorer_folder = '/path/to/exposome_explorer/data/folder/'
processed_data_folder = "/path/to/processed/data/folder/"

merged_databases = pd.read_csv(os.path.join(processed_data_folder, 'foodb_hmdb_markerdb_chebi.csv'), dtype={'drugbank_id': 'string', 'knapsack_id': 'string', 'wikipedia_id': 'string', 'biocyc_id': 'string', 'vmh_id': 'string', 'pdb_id': 'string', 'description': 'string', 'cas_number': 'string', 'kingdom': 'string', 'superclass': 'string', 'class': 'string', 'subclass': 'string', 'chebi_id': 'string', 'kegg_id': 'string','lipid_maps': 'string', 'meta_cyc': 'string', 'synonym': 'string', 'foodb_id': 'string', 'markerdb_id': 'string'})

exposome_explorer = pd.read_csv(os.path.join(exposome_explorer_folder, 'biomarkers.csv'), dtype={'ChEBI ID': 'string', 'PubChem ID': 'string'})

## Prepare Exposome Explorer Data

exposome_explorer_single = exposome_explorer[exposome_explorer.Level == 'Single']
exposome_explorer_single = exposome_explorer_single.drop(columns=['Level', 'Synonyms', 'No. of Publications', 'No. of Concentration values', 'No. of Reproducibility values', 'No. of Correlation values', 'No. of Metabolomic associations', 'No. of Microbiota associations', 'No. of Cancer associations', 'Average mass'])

exposome_explorer_single = exposome_explorer_single.rename(columns={'ID': 'exposome_explorer_id', 'Name': 'name_exposome_explorer', 'Classification': 'classification', 'Description': 'description_exposome_explorer', 'CAS Number': 'cas_number_exposome_explorer', 'PubChem ID': 'pubchem_compound_id_exposome_explorer', 'ChEBI ID': 'chebi_id_exposome_explorer', 'FooDB ID': 'foodb_id_exposome_explorer', 'HMDB ID': 'hmdb_id_exposome_explorer', 'SMILES': 'smiles_exposome_explorer', 'Formula': 'chemical_formula_exposome_explorer', 'InChI': 'inchi_exposome_explorer', 'InChIKey': 'inchikey_exposome_explorer', 'Mono. mass': 'mono_mass_exposome_explorer'})

## Merge

matched_exposome_eplorer_ids = set()

def match_exposome_explorer_via_chebi(row):
    chebi_id_match = exposome_explorer_single[exposome_explorer_single['chebi_id_exposome_explorer'] == row.chebi_id]
    if len(chebi_id_match) > 1:
        print('Error: more than one ChEBI match!')
    elif len(chebi_id_match) == 1:
        matched_exposome_eplorer_ids.add(list(chebi_id_match['exposome_explorer_id'])[0])
        return chebi_id_match.iloc[0]
    elif len(chebi_id_match) == 0:
        return pd.Series(None, index=exposome_explorer_single.columns)

def match_exposome_explorer_via_hmdb(row):
    if pd.isna(row["exposome_explorer_id"]):
        hmdb_id_match = exposome_explorer_single[exposome_explorer_single['hmdb_id_exposome_explorer'] == row.hmdb_id]
        if len(hmdb_id_match) > 1:
            print('Error: more than one HMDB match!')
        elif len(hmdb_id_match) == 1:
            ee_id = list(hmdb_id_match['exposome_explorer_id'])[0]
            if ee_id not in matched_exposome_eplorer_ids:
                matched_exposome_eplorer_ids.add(ee_id)
                return hmdb_id_match.iloc[0]
            else:
                return row[list(exposome_explorer_single.columns)]
        elif len(hmdb_id_match) == 0:
            return row[list(exposome_explorer_single.columns)]
    else: 
        return row[list(exposome_explorer_single.columns)]

def match_exposome_explorer_via_foodb(row):
    if pd.isna(row["exposome_explorer_id"]):
        foodb_id_match = exposome_explorer_single[exposome_explorer_single['foodb_id_exposome_explorer'] == row.foodb_id]
        if len(foodb_id_match) > 1:
            print('Error: more than one FooDB match!')
        elif len(foodb_id_match) == 1:
            ee_id = list(foodb_id_match['exposome_explorer_id'])[0]
            if ee_id not in matched_exposome_eplorer_ids:
                matched_exposome_eplorer_ids.add(ee_id)
                return foodb_id_match.iloc[0]
            else:
                return row[list(exposome_explorer_single.columns)]
        elif len(foodb_id_match) == 0:
            return row[list(exposome_explorer_single.columns)]
    else: 
        return row[list(exposome_explorer_single.columns)]

def match_exposome_explorer_via_name(row):
    if pd.isna(row["exposome_explorer_id"]):
        name_match = exposome_explorer_single[exposome_explorer_single['name_exposome_explorer'].str.lower() == row.name]
        if len(name_match) > 1:
            print('Error: more than one name match!')
        elif len(name_match) == 1:
            ee_id = list(name_match['exposome_explorer_id'])[0]
            if ee_id not in matched_exposome_eplorer_ids:
                matched_exposome_eplorer_ids.add(ee_id)
                return name_match.iloc[0]
            else:
                return row[list(exposome_explorer_single.columns)]
        elif len(name_match) == 0:
            return row[list(exposome_explorer_single.columns)]
    else: 
        return row[list(exposome_explorer_single.columns)]

merged_databases[exposome_explorer_single.columns] = merged_databases.apply(match_exposome_explorer_via_chebi, axis=1)
merged_databases[exposome_explorer_single.columns] = merged_databases.apply(match_exposome_explorer_via_hmdb, axis=1)
merged_databases[exposome_explorer_single.columns] = merged_databases.apply(match_exposome_explorer_via_foodb, axis=1)
merged_databases[exposome_explorer_single.columns] = merged_databases.apply(match_exposome_explorer_via_name, axis=1)

len(matched_exposome_eplorer_ids)

len(exposome_explorer_single)

unmatched_exposome_eplorer = exposome_explorer_single[~exposome_explorer_single.exposome_explorer_id.isin(matched_exposome_eplorer_ids)]
unmatched_exposome_eplorer = unmatched_exposome_eplorer.reindex(columns=merged_databases.columns)
unmatched_exposome_eplorer

all_data_merge = pd.concat([merged_databases, unmatched_exposome_eplorer])

all_data_merge.info()

## Clean up merged dataframe

all_data_merge = all_data_merge.drop(columns=['hmdb_id_exposome_explorer', 'foodb_id_exposome_explorer', 'chebi_id_exposome_explorer'])

columns_to_clean = ['name', 'description', 'cas_number', 'pubchem_compound_id', 'chemical_formula', 'mono_mass', 'inchi', 'inchikey', 'smiles']

for column in columns_to_clean:
    column_ee = column + '_exposome_explorer'

    all_data_merge[column] = all_data_merge.apply(lambda row: row[column] if pd.notna(row[column]) else row[column_ee], axis=1)
    all_data_merge = all_data_merge.drop(columns=[column_ee])

all_data_merge['name'] = all_data_merge['name'].apply(lambda x: x.lower())

all_data_merge.to_csv(os.path.join(processed_data_folder, 'foodb_hmdb_markerdb_chebi_exposomeexplorer.csv'), index=False)
