==> database_parsing/10_match_kegg.ipynb <==
*MeNu GUIDE*

# Match KEGG

import pandas as pd
import os

kegg_folder = '/path/to/kegg/data/folder/'
processed_data_folder = "/path/to/processed/data/folder/"

merged_data = pd.read_csv(os.path.join(processed_data_folder, 'foodb_hmdb_markerdb_chebi_exposomeexplorer.csv'), dtype={'drugbank_id': 'string', 'knapsack_id': 'string', 'wikipedia_id': 'string', 'biocyc_id': 'string', 'vmh_id': 'string', 'pdb_id': 'string', 'description': 'string', 'cas_number': 'string', 'kingdom': 'string', 'superclass': 'string', 'class': 'string', 'subclass': 'string', 'chebi_id': 'string', 'kegg_id': 'string','lipid_maps': 'string', 'meta_cyc': 'string', 'synonym': 'string', 'foodb_id': 'string', 'markerdb_id': 'string', 'classification': 'string'})
kegg_compounds = pd.read_csv(os.path.join(kegg_folder, 'kegg_compounds_processed.csv'))

## Prepare kegg dataframe

kegg_compounds = kegg_compounds.drop(columns=['molecular_weight', 'NIKKAJI', '3DMET', 'Drug group', 'PubChem_subtance_id', 'PDB-CCD', 'pubchem_name'])

kegg_compounds = kegg_compounds.rename(columns={'entry_id': 'kegg_id', 'formula': 'chemical_formula', 'exact_mass': 'mono_mass', 'inchi_key': 'inchikey', 'LIPIDMAPS': 'lipid_maps', 'KNApSAcK': 'knapsack_id', 'ChEBI': 'chebi_id', 'CAS': 'cas_number'})

## Prepare merged dataframe

kegg_compounds['name'] = kegg_compounds.name.str.lower()
kegg_compounds['name'] = kegg_compounds.name.str.strip()
kegg_name_dict = kegg_compounds[['name', 'kegg_id']].set_index('name').to_dict()
kegg_name_dict = kegg_name_dict['kegg_id']
merged_data['kegg_id'] = merged_data.apply(lambda row: (kegg_name_dict[row['name']] if row['name'] in kegg_name_dict else float('NaN')) if pd.isna(row.kegg_id) else row.kegg_id, axis=1)

## Check duplicate KEGG IDs in merged dataset

merged_data[(merged_data.kegg_id.duplicated()) & (merged_data.kegg_id.notna())][['name', 'kegg_id']]

merged_data[merged_data.kegg_id == "C00596"][['name', 'hmdb_id', 'foodb_id', 'chebi_id', 'kegg_id']]

kegg_compounds[kegg_compounds.kegg_id == "C00596"]

Seems like there are multiple correct chebi_ids for one kegg_entry, probably a hierarchy & stereoisomer issue. So I can't really remove any identifier and I will just leave duplicated kegg_ids. 

## Merge datasets

merged_data_kegg = merged_data.merge(kegg_compounds, on='kegg_id', how='outer', suffixes=['', '_kegg'])

merged_data_kegg[(merged_data_kegg.kegg_id.notna()) & (merged_data_kegg.name_kegg.isna())][['name', 'hmdb_id', 'foodb_id', 'chebi_id', 'kegg_id']]

# These compounds don't seem to exist in the KEGG database, so I am going to overwrite them
merged_data_kegg['kegg_id'] = merged_data_kegg.apply(lambda row: float('NaN') if (pd.notna(row.kegg_id) & pd.isna(row.name_kegg)) else row.kegg_id, axis=1)

## Clean up data merge

merged_data_kegg = merged_data_kegg.drop(columns=['chebi_id_kegg'])

columns_to_clean = ['name', 'lipid_maps', 'knapsack_id', 'cas_number', 'pubchem_compound_id', 'chemical_formula', 'mono_mass', 'inchi', 'inchikey', 'smiles']

for column in columns_to_clean:
    column_kegg = column + '_kegg'

    merged_data_kegg[column] = merged_data_kegg.apply(lambda row: row[column] if pd.notna(row[column]) else row[column_kegg], axis=1)
    merged_data_kegg = merged_data_kegg.drop(columns=[column_kegg])

merged_data_kegg.to_csv(os.path.join(processed_data_folder, "foodb_hmdb_markerdb_chebi_exposomeexplorer_kegg.csv"), index=False)
