==> database_parsing/11_match_vmh.ipynb <==
*MeNu GUIDE*

# Match VMH/Recon3

import pandas as pd
import os

vmh_folder = '/path/to/vmh/data/folder/'
processed_data_folder = "/path/to/processed/data/folder/"

merged_data = pd.read_csv(os.path.join(processed_data_folder, "foodb_hmdb_markerdb_chebi_exposomeexplorer_kegg.csv"), dtype={'drugbank_id': 'string', 'knapsack_id': 'string', 'wikipedia_id': 'string', 'biocyc_id': 'string', 'vmh_id': 'string', 'pdb_id': 'string', 'description': 'string', 'cas_number': 'string', 'kingdom': 'string', 'superclass': 'string', 'class': 'string', 'subclass': 'string', 'chebi_id': 'string', 'kegg_id': 'string','lipid_maps': 'string', 'meta_cyc': 'string', 'synonym': 'string', 'foodb_id': 'string', 'markerdb_id': 'string', 'classification': 'string'})

## Clean up VMH ID duplicates

merged_data[(merged_data.vmh_id.duplicated()) & (merged_data.vmh_id.notna())][['hmdb_id', 'foodb_id', 'chebi_id', 'kegg_id', 'vmh_id', 'name']]

merged_data.loc[10, 'vmh_id'] = float('NaN') # HMDB0000217 
merged_data.loc[794, 'vmh_id'] = float('NaN') # HMDB0242149
merged_data.loc[1857, 'vmh_id'] = float('NaN') # HMDB0302754
merged_data.loc[94341, 'vmh_id'] = float('NaN') # HMDB0012515
merged_data.loc[353289, 'vmh_id'] = float('NaN') # HMDB0254327

## Match with VMH metabolites

vmh_metabolites = pd.read_csv(os.path.join(vmh_folder, "vmh_metabolites_all.csv"), dtype={'cheBlId': 'string'})

I have the following identifiers to merge with:
* iupac
* keggId
* pubChemId
* cheBlId
* inchiString
* inchiKey
* hmdb
* food_db

$\rightarrow$ start with hmdb, then foodb, then chebi

### Clean up VMH dataframe

vmh_metabolites['formula'] = vmh_metabolites.apply(lambda row: row.chargedFormula if pd.notna(row.chargedFormula) else row.neutralFormula, axis=1)

vmh_metabolites = vmh_metabolites.drop(columns=['Unnamed: 0', 'index', 'createdDate', 'updatedDate', 'lmId', 'ehmnId', 'hepatonetId', 'metanetx', 'seed', 'pdmapName', 'reconMap', 'golgimap', 'lysosomemap', 'mitochondrionmap', 'nucleusmap', 'reticulummap', 'peroxisomemap', 'epa_id', 'echa_id', 'iuphar_id', 'fda_id', 'mesh_id', 'chodb_id', 'isHuman', 'isMicrobe', 'charge', 'chargedFormula', 'neutralFormula', 'synonyms', 'met_id', 'miriam', 'avgmolweight', 'biggId', 'chembl'])

merged_data.columns

vmh_metabolites = vmh_metabolites.rename(columns={'abbreviation': 'vmh_id', 'fullName': 'name', 'monoisotopicweight': 'mono_mass', 'keggId': 'kegg_id', 'pubChemId': 'pubchem_compound_id', 'cheBlId': 'chebi_id', 'inchiString': 'inchi', 'inchiKey': 'inchikey', 'hmdb': 'hmdb_id', 'reconMap3': 'recon3', 'food_db': 'foodb_id', 'chemspider': 'chemspider_id', 'biocyc': 'biocyc_id', 'wikipedia': 'wikipedia_id', 'drugbank': 'drugbank_id', 'knapsack': 'knapsack_id', 'phenolExplorer': 'phenol_explorer_compound_id', 'metlin': 'metlin_id', 'casRegistry': 'cas_number', 'formula': 'chemical_formula', 'smile': 'smiles'})
vmh_metabolites['name'] = vmh_metabolites['name'].str.lower()

vmh_metabolites_hmdb = vmh_metabolites[vmh_metabolites.hmdb_id.notna()][['vmh_id', 'hmdb_id']]
vmh_metabolites_foodb = vmh_metabolites[vmh_metabolites.foodb_id.notna()][['vmh_id', 'foodb_id']]
vmh_metabolites_chebi = vmh_metabolites[vmh_metabolites.chebi_id.notna()][['vmh_id', 'chebi_id']]
vmh_metabolites_name = vmh_metabolites[vmh_metabolites.name.notna()][['vmh_id', 'name']]

vmh_abbreviations_hmdb = dict(zip(vmh_metabolites_hmdb.hmdb_id, vmh_metabolites_hmdb.vmh_id))
vmh_abbreviations_foodb = dict(zip(vmh_metabolites_foodb.foodb_id, vmh_metabolites_foodb.vmh_id))
vmh_abbreviations_chebi = dict(zip(vmh_metabolites_chebi.chebi_id, vmh_metabolites_chebi.vmh_id))
vmh_abbreviations_name = dict(zip(vmh_metabolites_name.name, vmh_metabolites_name.vmh_id))

### Match as many VMH metabolites as possible

merged_data['vmh_id'] = merged_data.vmh_id.str.lower()
vmh_metabolites['vmh_id'] = vmh_metabolites.vmh_id.str.lower()

def match_vmh(row):
    hmdb_match = vmh_abbreviations_hmdb[row.hmdb_id] if (pd.notna(row.hmdb_id) and (row.hmdb_id in vmh_abbreviations_hmdb.keys())) else None
    foodb_match = vmh_abbreviations_foodb[row.foodb_id] if (pd.notna(row.foodb_id) and (row.foodb_id in vmh_abbreviations_foodb.keys())) else None
    chebi_match = vmh_abbreviations_chebi[row.chebi_id] if (pd.notna(row.chebi_id) and (row.chebi_id in vmh_abbreviations_chebi.keys())) else None
    name_match = vmh_abbreviations_name[row.name] if (pd.notna(row.name) and (row.name in vmh_abbreviations_name.keys())) else None

    if pd.notna(row.vmh_id):
        return row.vmh_id.lower()
    elif pd.notna(name_match):
        return name_match.lower()
    elif pd.notna(hmdb_match):
        return hmdb_match.lower()
    elif pd.notna(foodb_match):
        return foodb_match.lower()
    elif pd.notna(chebi_match):
        return chebi_match.lower()
    else:
        return None

print(merged_data.vmh_id.nunique())
print(merged_data.vmh_id.count())

merged_data['vmh_id'] = merged_data.apply(match_vmh, axis=1)

print(merged_data.vmh_id.nunique())
print(merged_data.vmh_id.count())

### Merge dataframes

merge_temp = merged_data.merge(vmh_metabolites, how='outer', on='vmh_id', suffixes=['', '_vmh'])

merge_temp = merge_temp.drop(columns=['hmdb_id_vmh', 'chebi_id_vmh', 'foodb_id_vmh'])

merge_temp.columns

columns_to_clean = ['name', 'description', 'iupac', 'kegg_id', 'drugbank_id', 'cas_number', 'pubchem_compound_id', 'chemspider_id', 'biocyc_id', 'wikipedia_id', 'knapsack_id', 'phenol_explorer_compound_id', 'metlin_id', 'chemical_formula', 'mono_mass', 'inchi', 'inchikey', 'smiles']

for column in columns_to_clean:
    column_vmh = column + '_vmh'

    merge_temp[column] = merge_temp.apply(lambda row: row[column] if pd.notna(row[column]) else row[column_vmh], axis=1)
    merge_temp = merge_temp.drop(columns=[column_vmh])

merge_temp.columns

merge_temp.to_csv(os.path.join(processed_data_folder, "metabolites_all_databases_merged.csv"), index=False)
