==> database_parsing/12_clean_up_final_metabolite_dataframe.ipynb <==
*MeNu GUIDE*

# Clean up final metabolite dataframe

import os
import pandas as pd

processed_data_folder = "/path/to/processed/data/folder/"

merged_data = pd.read_csv(os.path.join(processed_data_folder, "metabolites_all_databases_merged.csv"), dtype={'drugbank_id': 'string', 'knapsack_id': 'string', 'wikipedia_id': 'string', 'biocyc_id': 'string', 'vmh_id': 'string', 'pdb_id': 'string', 'description': 'string', 'cas_number': 'string', 'kingdom': 'string', 'superclass': 'string', 'class': 'string', 'subclass': 'string', 'chebi_id': 'string', 'kegg_id': 'string','lipid_maps': 'string', 'meta_cyc': 'string', 'synonym': 'string', 'foodb_id': 'string', 'markerdb_id': 'string', 'classification': 'string', 'recon3': 'string'})

## Inchi Duplicates

duplicates_inchi = set(merged_data[merged_data.inchi.duplicated() & merged_data.inchi.notna()].inchi.unique())
merged_inchi_duplicated = merged_data[merged_data.inchi.isin(duplicates_inchi)].sort_values(by='inchi')

merged_inchi_duplicated[['name', 'hmdb_id', 'foodb_id', 'chebi_id', 'kegg_id', 'vmh_id', 'exposome_explorer_id', 'markerdb_id', 'inchi']]

def find_conflicts(rows):
    row_columns = rows.columns
    
    merged_row = {}
    
    merge_conflict = False
    
    for index, row in rows.iterrows():
        #print(f"{index}: {row['name']}")
        for column in row_columns:
            if column not in merged_row:
               merged_row[column] = row[column]
            elif pd.isna(merged_row[column]):
                merged_row[column] = row[column]
            elif pd.notna(row[column]) and merged_row[column] != row[column]:
                if column in ['hmdb_id', 'chebi_id', 'foodb_id', 'markerdb_id', 'kegg_id', 'vmh_id', 'exposome_explorer_id']:
                    #print(f'Merge conflict: {merged_row[column]} != {row[column]} for {column}')
                    merge_conflict = True
                elif column == 'name':
                    if len(row[column]) < len(merged_row[column]):
                        merged_row[column] = row[column]
                    
    return merged_row, merge_conflict

solvable_duplicates = 0
merge_conflict_count = 0

merged_rows = []

for inchi in duplicates_inchi:
    merged_row, merge_conflict = find_conflicts(merged_data[merged_data.inchi == inchi])
    if merge_conflict:
        merge_conflict_count += 1
    else:
        solvable_duplicates += 1
        merged_rows.append(merged_row)

print(f"Solvable {solvable_duplicates} vs. not solvable {merge_conflict_count}")

merged_rows_df = pd.DataFrame(merged_rows)

merged_rows_inchis = set(merged_rows_df.inchi.unique())
merged_data = merged_data[~merged_data.inchi.isin(merged_rows_inchis)]
merged_data_inchi_clean = pd.concat([merged_data, merged_rows_df])

## Clean up name duplicates

duplicated_names = set(merged_data_inchi_clean[merged_data_inchi_clean.name.duplicated()].name.unique())
merged_name_duplicated_inchi_clean = merged_data_inchi_clean[merged_data_inchi_clean.name.isin(duplicated_names)].sort_values(by='name')

merged_name_duplicated_inchi_clean

solvable_duplicates_name = 0
merge_conflict_count_name = 0

merged_rows_name = []

for name in duplicated_names:
    merged_row_name, merge_conflict_name = find_conflicts(merged_data_inchi_clean[merged_data_inchi_clean.name == name])
    if merge_conflict_name:
        merge_conflict_count_name += 1
    else:
        solvable_duplicates_name += 1
        merged_rows_name.append(merged_row_name)

print(f"Solvable {solvable_duplicates_name} vs. not solvable {merge_conflict_count_name} of total {len(duplicated_names)}")

merged_rows_name_df = pd.DataFrame(merged_rows_name)
merged_rows_names = set(merged_rows_name_df.name.unique())
merged_data_inchi_clean = merged_data_inchi_clean[~merged_data_inchi_clean.name.isin(merged_rows_names)]
merged_data_inchi_names_clean = pd.concat([merged_data_inchi_clean, merged_rows_name_df])

merged_data_inchi_names_clean

merged_data_inchi_names_clean.info()

merged_data_inchi_names_clean.to_csv(os.path.join(processed_data_folder, 'compounds_all_databases_merged.csv'), index=False)

set(merged_data_inchi_names_clean.columns) - {'name', 'inchi', 'inchikey', 'chemical_formula', 'smiles', 'mono_mass', 'cas_number', 'description', 'iupac'}
