==> database_parsing/13_parse_foodb_foods.ipynb <==
Nutrition-Disease Knowledge Graph
# FooDB Parsing

import os.path
import pandas as pd

## Food Table

foodb_folder = "/path/to/downloaded/FooDB/data/"
processed_data_folder = "/path/to/processed/data/folder/"

food = pd.read_csv(os.path.join(foodb_folder, "Food.csv"), dtype={'ncbi_taxonomy_id': 'string', 'itis_id': 'string', 'legacy_id': 'string'})

food = food[['id', 'name', 'name_scientific', 'description', 'food_type', 'ncbi_taxonomy_id', 'public_id']]

food

food.to_csv(os.path.join(foodb_folder, 'foods_to_map.csv'), index=False)

## Food Contents
### Process content table

content_table = pd.read_csv(os.path.join(foodb_folder, "Content.csv"), dtype={'source_id': int})

print(len(content_table))
content_table = content_table[content_table.source_type == 'Compound']
print(len(content_table))

print(len(content_table))
content_table = content_table[content_table.citation_type.isin(['DATABASE', 'EXPERIMENTAL', 'ARTICLE', 'TEXTBOOK'])]
print(len(content_table))

print(len(content_table))
content_table = content_table[content_table.orig_unit.notna()]
print(len(content_table))

print(len(content_table))
content_table = content_table[content_table.orig_content.notna()]
print(len(content_table))

print(len(content_table))
content_table = content_table[content_table.orig_content != 0]
print(len(content_table))

content_table = content_table[['id', 'source_id', 'orig_source_name', 'food_id', 'orig_food_common_name', 'orig_content', 'orig_min', 'orig_max', 'orig_unit', 'orig_citation', 'citation', 'citation_type']]
content_table = content_table.rename(columns={'source_id': 'compound_id', 'orig_food_common_name': 'food_name', 'orig_food_scientific_name': 'food_scientific_name', 'orig_unit': 'unit', 'orig_content': 'content', 'orig_min': 'min_content', 'orig_max': 'max_content', 'orig_citation': 'citation_link', 'orig_source_name': 'compound', })

content_table

### Load compounds table 

# Load merged database compounds file
compounds = pd.read_csv(os.path.join(processed_data_folder, "compounds_all_databases_merged_ids.csv"), dtype={'vmh_id': 'string', 'chebi_id': 'string', 'kegg_id': 'string', 'foodb_id': 'string', 'markerdb_id': 'string', 'exposome_explorer_id': 'string'})

compounds_temp = compounds[compounds.foodb_id_internal.notna()][['name', 'foodb_id', 'foodb_id_internal', 'menuguide_id']]
compounds_temp['foodb_id_internal'] = compounds_temp.foodb_id_internal.astype(int)

compounds_temp

### Merge content and compounds to add MeNuGUIDE identifier

content_table_merge = content_table.merge(compounds_temp, left_on='compound_id', right_on='foodb_id_internal', how='inner')

content_table_merge['citation'] = content_table_merge.apply(lambda row: f"{row['citation']}:{row['citation_link']}" if pd.notna(row['citation_link']) else row['citation'], axis=1)
content_table_merge['citation'] = content_table_merge['citation'].apply(lambda x: f"PMID:{x}" if x.isnumeric() else x)

content_table_merge = content_table_merge[['food_id', 'food_name', 'name', 'foodb_id', 'foodb_id_internal', 'menuguide_id', 'content', 'min_content', 'max_content', 'unit', 'citation']]

content_table_merge.to_csv(os.path.join(foodb_folder, 'foodb_metabolite_food_relations.csv'), index=False)
