==> database_parsing/14_clean_up_markerdb_tables.ipynb <==
*MeNu GUIDE*

# Clean up MarkerDB tables

import pandas as pd
import os

markerdb_folder = "/path/to/markerdb/data/folder/"
processed_data_folder = "/path/to/processed/data/folder/"

## Chemicals

all_chemicals = pd.read_csv(os.path.join(markerdb_folder, "all_chemicals.tsv"), header=None, sep='\t')

all_chemicals = all_chemicals.rename(columns={0: 'markerdb_id', 1: 'name_markerdb', 2: 'condition_id', 3: 'condition', 4: 'cohort', 5: 'sex', 6: 'sample_type', 7: 'concentration', 8: 'unit', 9: 'notes', 10: 'reference_id', 11: 'reference_name', 12: 'pubmed_id'})

all_chemicals

all_chemicals.condition.unique()

all_chemicals.condition.nunique()

all_chemicals

all_chemicals.to_csv(os.path.join(processed_data_folder, 'markerdb_chemical_associations.csv'), index=False)

## Proteins

all_proteins = pd.read_csv(os.path.join(markerdb_folder, "all_proteins.tsv"), header=None, sep='\t')
all_proteins    

all_proteins = all_proteins.rename(columns={0: 'markerdb_id', 1: 'name_markerdb', 2: 'condition_id', 3: 'condition', 4: 'cohort', 5: 'sex', 6: 'sample_type', 7: 'concentration', 8: 'unit', 9: 'notes', 10: 'reference_name', 11: 'pubmed_id'})
all_proteins

all_proteins.to_csv(os.path.join(processed_data_folder, 'all_protein_associations_markerdb.csv'), index=False)

## Genes

all_genes = pd.read_csv(os.path.join(markerdb_folder, "all_genetics.tsv"), header=None, sep='\t')
all_genes

all_genes = all_genes.drop(columns=[5, 6, 7, 8])

all_genes = all_genes.rename(columns={0: 'markerdb_id', 1: 'gene_symbol', 2: 'snp_id', 3: 'condition', 4: 'clinical_significance'})
all_genes

all_genes.to_csv(os.path.join(processed_data_folder, 'all_gene_associations_markerdb.csv'), index=False)
