==> database_parsing/15_match_genes_enzymes_proteins.ipynb <==
*MeNu GUIDE*

# Match Genes and Enzymes

import os
import pandas as pd

kegg_folder = '/path/to/kegg/data/folder/'
vmh_folder = '/path/to/vmh/data/folder/'
processed_data_folder = "/path/to/processed/data/folder/"

kegg_proteins = pd.read_csv(os.path.join(kegg_folder, "kegg_enzymes.csv"))
kegg_genes = pd.read_csv(os.path.join(kegg_folder, "kegg_enzyme_genes.csv"))
vmh_genes = pd.read_csv(os.path.join(vmh_folder, "vmh_human_genes.csv"))

## Clean up KEGG genes table

kegg_genes = kegg_genes[kegg_genes.gene.str.contains('\(')]
kegg_genes.loc[:, 'gene_name'] = kegg_genes.gene.apply(lambda x: x.split('(')[1].rstrip(')'))
kegg_genes.loc[:, 'kegg_gene_id'] = kegg_genes.gene.apply(lambda x: f"hsa:{x.split('(')[0]}")
kegg_genes = kegg_genes.drop(columns='gene')
kegg_genes = kegg_genes.rename(columns={'enzyme_id': 'kegg_enzyme_id'})
kegg_genes

## Merge KEGG genes and proteins tables

kegg_proteins = kegg_proteins.merge(kegg_genes, how='inner', on='kegg_enzyme_id')
kegg_proteins = kegg_proteins.rename(columns={'gene_name': 'symbol'})
kegg_proteins

## Match VMH & KEGG

vmh_genes = vmh_genes.drop(columns=['Unnamed: 0', 'genes_id', 'abbreviation'])

kegg_vmh_merge = vmh_genes.merge(kegg_proteins, how='outer', on='symbol')
print(f"Number of unmatched VMH genes: {len(kegg_vmh_merge[kegg_vmh_merge.gene_number.isna()])}")
print(f"Number of unmatched KEGG genes: {len(kegg_vmh_merge[kegg_vmh_merge.kegg_enzyme_id.isna()])}")
print(f"Number of matched genes: {len(kegg_vmh_merge[(kegg_vmh_merge.kegg_enzyme_id.notna()) & (kegg_vmh_merge.gene_number.notna())])}")

kegg_vmh_merge.to_csv(os.path.join(processed_data_folder, "proteins_genes_kegg_vmh.csv"), index=False)

kegg_vmh_merge_genes = kegg_vmh_merge[['symbol', 'gene_number', 'chromosome', 'description', 'ensembl_gene', 'chebl_id', 'uniprot_gname', 'kegg_gene_id']]
kegg_vmh_merge_genes.to_csv(os.path.join(processed_data_folder, "genes_kegg_vmh.csv"), index=False)

kegg_vmh_merge_genes_to_enzymes = kegg_vmh_merge[['symbol', 'kegg_enzyme_id']]
kegg_vmh_merge_genes_to_enzymes = kegg_vmh_merge_genes_to_enzymes[kegg_vmh_merge_genes_to_enzymes.kegg_enzyme_id.notna()]
kegg_vmh_merge_genes_to_enzymes = kegg_vmh_merge_genes_to_enzymes.drop_duplicates()
kegg_vmh_merge_genes_to_enzymes.to_csv(os.path.join(processed_data_folder, "kegg_enzyme_gene_symbols.csv"), index=False)

kegg_vmh_merge_genes_to_enzymes[kegg_vmh_merge_genes_to_enzymes.kegg_enzyme_id == 'EC 1.1.1.1']
