==> database_parsing/16_match_reactions.ipynb <==
*MeNu GUIDE*

# Match Reactions

import pandas as pd
import ast
import os

## Prepare VMH data

vmh_folder = '/path/to/vmh/data/folder/'
kegg_folder = '/path/to/kegg/data/folder/'
processed_data_folder = "/path/to/processed/data/folder/"

reactions = pd.read_csv(os.path.join(vmh_folder, "vmh_reactions.csv"))
reactions

reaction_substrates = reactions[['abbreviation', 'substrates']]
reaction_products = reactions[['abbreviation', 'products']] 

reaction_substrates.loc[:, 'substrates'] = reaction_substrates['substrates'].apply(lambda x: ast.literal_eval(x))
reaction_products.loc[:, 'products'] = reaction_products['products'].apply(lambda x: ast.literal_eval(x))

substrates_parsed = [['substrate', 'reaction_abbreviation']]

def separate_substrates(row):
    reaction_id = row.abbreviation
    substrates = row.substrates
    reaction_substrates = list(map(lambda e: (e, reaction_id), substrates))
    substrates_parsed.extend(reaction_substrates)

reaction_substrates.apply(separate_substrates, axis=1)

substrates_parsed_df = pd.DataFrame(substrates_parsed[1:], columns=substrates_parsed[0])
substrates_parsed_df

products_parsed = [['product', 'reaction_abbreviation']]

def separate_products(row):
    reaction_id = row.abbreviation
    products = row.products
    reaction_products = list(map(lambda e: (e, reaction_id), products))
    products_parsed.extend(reaction_products)

reaction_products.apply(separate_products, axis=1)

products_parsed_df = pd.DataFrame(products_parsed[1:], columns=products_parsed[0])
products_parsed_df

reactions = reactions.drop(columns=['Unnamed: 0', 'isHuman', 'isMicrobe', 'substrates', 'products'])

reactions.to_csv(os.path.join(vmh_folder, "vmh_reactions.csv"), index=False)
substrates_parsed_df.to_csv(os.path.join(vmh_folder, "vmh_reaction_substrates.csv"), index=False)
products_parsed_df.to_csv(os.path.join(vmh_folder, "vmh_reaction_products.csv"), index=False)

## Match reactions

compound_ids = pd.read_csv(os.path.join(processed_data_folder, "compounds_all_databases_merged_ids.csv"), dtype=str)

kegg_vmh_compounds_match = compound_ids[['menuguide_id', 'kegg_id', 'vmh_id']]
kegg_vmh_compounds_match = kegg_vmh_compounds_match[kegg_vmh_compounds_match.kegg_id.notna() & kegg_vmh_compounds_match.vmh_id.notna()]
kegg_vmh_compounds_match

### KEGG

kegg_reactions = pd.read_csv(os.path.join(kegg_folder, "kegg_reactions.csv"))
kegg_substrates = pd.read_csv(os.path.join(kegg_folder, "kegg_reaction_substrates.csv"))
kegg_products = pd.read_csv(os.path.join(kegg_folder, "kegg_reaction_products.csv"))

kegg_substrates = kegg_substrates.merge(kegg_vmh_compounds_match, how='left', left_on='substrate', right_on='kegg_id')[['vmh_id', 'substrate', 'reaction_id']]
kegg_products = kegg_products.merge(kegg_vmh_compounds_match, how='left', left_on='product', right_on='kegg_id')[['vmh_id', 'product', 'reaction_id']]

kegg_substrates.loc[kegg_substrates.vmh_id.isna(),'vmh_id'] = 'exclude'
kegg_products.loc[kegg_products.vmh_id.isna(),'vmh_id'] = 'exclude'

kegg_reaction_list = list(kegg_reactions.reaction_id.unique())

kegg_reaction_dict = {}

for reaction in kegg_reaction_list:
    substrates = set(kegg_substrates[kegg_substrates.reaction_id == reaction].vmh_id)
    products = set(kegg_products[kegg_products.reaction_id == reaction].vmh_id)

    kegg_reaction_dict[reaction] = {'products': products, 'substrates': substrates}

### VMH

vmh_reactions = pd.read_csv(os.path.join(vmh_folder, "vmh_reactions.csv"))
vmh_substrates = pd.read_csv(os.path.join(vmh_folder, "vmh_reaction_substrates.csv"))
vmh_products = pd.read_csv(os.path.join(vmh_folder, "vmh_reaction_products.csv"))

vmh_reaction_list = list(vmh_reactions.abbreviation.unique())

hits = 0

reaction_matches = []

for reaction in vmh_reaction_list:
    vmh_substrates_temp = set(vmh_substrates[vmh_substrates.reaction_abbreviation == reaction].substrate)
    vmh_products_temp = set(vmh_products[vmh_products.reaction_abbreviation == reaction]['product'])

    for kegg_reaction_id, kegg_reaction_compounds in kegg_reaction_dict.items():
        kegg_substrates_temp = kegg_reaction_compounds['substrates']
        kegg_products_temp = kegg_reaction_compounds['products']

        same_substrates = vmh_substrates_temp == kegg_substrates_temp
        same_products = vmh_products_temp == kegg_products_temp

        same_substrates_reverse = vmh_substrates_temp == kegg_products_temp
        same_products_reverse = vmh_products_temp == kegg_substrates_temp

        if same_substrates and same_products:
            print(f'Found an exact hit! Number of hits: {hits}')
            reaction_matches.append({'vmh_reaction_abbreviation': reaction, 'kegg_reaction_id': kegg_reaction_id})
            hits += 1
        elif same_substrates_reverse and same_products_reverse:
            print(f'Found a reverse hit! Number of hits: {hits}')
            reaction_matches.append({'vmh_reaction_abbreviation': reaction, 'kegg_reaction_id': kegg_reaction_id})
            hits += 1

reaction_matches_df = pd.DataFrame(reaction_matches)
reaction_matches_df

reaction_matches_df[reaction_matches_df.vmh_reaction_abbreviation.duplicated()]

reaction_matches_df[reaction_matches_df.vmh_reaction_abbreviation == 'FBA']

vmh_reactions[vmh_reactions.abbreviation == 'FBA']

vmh_reactions_expanded = vmh_reactions.merge(reaction_matches_df, how='left', left_on='abbreviation', right_on='vmh_reaction_abbreviation')
vmh_reactions_expanded['keggId'] = vmh_reactions_expanded.apply(lambda row: row.keggId if pd.notna(row.keggId) else row.kegg_reaction_id, axis=1)
vmh_reactions_expanded = vmh_reactions_expanded.drop(columns=['kegg_reaction_id', 'vmh_reaction_abbreviation', 'rxn_id'])
vmh_reactions_expanded = vmh_reactions_expanded.rename(columns={'abbreviation': 'vmh_reaction_id', 'keggId': 'kegg_reaction_id'})

kegg_reactions = kegg_reactions.rename(columns={'reaction_id': 'kegg_reaction_id', 'reaction_name': 'kegg_reaction_name'})

vmh_kegg_reactions_merged = vmh_reactions_expanded.merge(kegg_reactions, how='outer', on='kegg_reaction_id')
vmh_kegg_reactions_merged

vmh_kegg_reactions_merged['reaction_identifier'] = vmh_kegg_reactions_merged.apply(lambda row: row.vmh_reaction_id if pd.notna(row.vmh_reaction_id) else row.kegg_reaction_id, axis=1)

duplicated_identifiers = list(vmh_kegg_reactions_merged[vmh_kegg_reactions_merged['reaction_identifier'].duplicated()].reaction_identifier.unique())

vmh_kegg_reactions_merged[vmh_kegg_reactions_merged['reaction_identifier'].isin(duplicated_identifiers)]

vmh_kegg_reactions_merged = vmh_kegg_reactions_merged.drop_duplicates()

vmh_kegg_reactions_merged = vmh_kegg_reactions_merged[['reaction_identifier', 'vmh_reaction_id', 'kegg_reaction_id', 'kegg_reaction_name', 'description', 'formula', 'reversible', 'notes','ref', 'ecnumber','rhea', 'reconMap3', 'organellemap', 'chodb_id', 'subsystem']]

vmh_kegg_reactions_merged = vmh_kegg_reactions_merged.reset_index(names='menuguide_id')
vmh_kegg_reactions_merged.loc[:, 'menuguide_id'] = vmh_kegg_reactions_merged.menuguide_id.apply(lambda x: f"reaction_{x}")

vmh_kegg_reactions_merged

vmh_kegg_reactions_merged.to_csv(os.path.join(processed_data_folder, "reactions_kegg_vmh.csv"), index=False)
